Added different json serialize engines support and swashbuckle support (wip)

+++ tests/Patcheetah.Tests/Models/Abstract/IUser.cs
@@ -6,7 +6,7 @@ namespace Patcheetah.Tests.Models.Abstract
     {
         public string Id { get; set; }
 
-        public long Age { get; set; }
+        public int Age { get; set; }
 
         public string Username { get; set; }
