+++ tests/Patcheetah.Tests/Models/WithAttributes/User.cs
@@ -10,7 +10,7 @@ namespace Patcheetah.Tests.Models.WithAttributes
         [PatchingKey]
         public string Id { get; set; }
 
-        public long Age { get; set; }
+        public int Age { get; set; }
 
         public string Username { get; set; }
